+++ phoenicis-javafx/src/main/java/org/phoenicis/javafx/themes/Theme.java
@@ -1,4 +1,4 @@
-package org.phoenicis.javafx.views.common.themes;
+package org.phoenicis.javafx.themes;
 
 import java.net.URI;
