+++ phoenicis-javafx/src/main/java/org/phoenicis/javafx/themes/ThemeConfiguration.java
@@ -16,9 +16,8 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-package org.phoenicis.javafx.views.common;
+package org.phoenicis.javafx.themes;
 
-import org.phoenicis.javafx.views.common.themes.Themes;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
