+++ phoenicis-javafx/src/main/java/org/phoenicis/javafx/views/mainwindow/settings/SettingsView.java
@@ -28,9 +28,9 @@ import org.phoenicis.javafx.components.setting.control.*;
 import org.phoenicis.javafx.components.setting.utils.ApplicationBuildInformation;
 import org.phoenicis.javafx.components.setting.utils.SettingsSidebarItem;
 import org.phoenicis.javafx.settings.JavaFxSettingsManager;
-import org.phoenicis.javafx.views.common.ThemeManager;
-import org.phoenicis.javafx.views.common.themes.Theme;
-import org.phoenicis.javafx.views.common.themes.Themes;
+import org.phoenicis.javafx.themes.ThemeManager;
+import org.phoenicis.javafx.themes.Theme;
+import org.phoenicis.javafx.themes.Themes;
 import org.phoenicis.javafx.views.mainwindow.ui.MainWindowView;
 import org.phoenicis.repository.RepositoryLocationLoader;
 import org.phoenicis.repository.RepositoryManager;
