+++ phoenicis-javafx/src/main/java/org/phoenicis/javafx/views/mainwindow/ui/MainWindowView.java
@@ -23,7 +23,7 @@ import javafx.scene.control.ProgressIndicator;
 import javafx.scene.control.Tab;
 import javafx.scene.layout.BorderPane;
 import javafx.scene.layout.HBox;
-import org.phoenicis.javafx.views.common.ThemeManager;
+import org.phoenicis.javafx.themes.ThemeManager;
 
 import java.util.Optional;
